pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "thegooseforce"
version = "0.1.3"
description = "TheGooseForce - A Salesforce integration extension for Goose"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["goose", "salesforce", "mcp"]
dependencies = [
    "mcp",
    "simple-salesforce",
    "python-dotenv",
]
license = { text = "MIT" }
authors = [
    { name = "Tristan Nohrer" },
]

[project.urls]
Homepage = "https://github.com/tristannohrer/TheGooseForce"

[project.scripts]
thegooseforce = "thegooseforce.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/thegooseforce"]